naive

+++ BoundedGirth/naive_edge/enum.cpp
@@ -70,19 +70,19 @@ bool Girth(Graph &g){
 
 
 std::vector<bigint> ELGMain(Graph &g){
-  int m = 0;
+  long long int m = 0;
   for (int i = 0; i < g.size(); i++) {
     m += g[i].size();
   }
-  m /= 2;
+  m /= 2LL;
   std::cout << m << std::endl;
   ans.resize(m + 1, 0);
-  for (int i = 1; i < (1<<m); i++) {
+  for (int i = 1; i < (1LL<<m); i++) {
     int size = 0;
-    if(i == (1<<m)/16)std::cout << "1/16" << std::endl;
-    if(i == (1<<m)/8)std::cout << "1/8" << std::endl;
-    if(i == (1<<m)/4)std::cout << "1/4" << std::endl;
-    if(i == (1<<m)/2)std::cout << "1/2" << std::endl;
+    if(i == (1LL<<m)/16)std::cout << "1/16" << std::endl;
+    if(i == (1LL<<m)/8)std::cout << "1/8" << std::endl;
+    if(i == (1LL<<m)/4)std::cout << "1/4" << std::endl;
+    if(i == (1LL<<m)/2)std::cout << "1/2" << std::endl;
     for (int j = 0; j < m; j++) {
       pii p = g.edge_id[j].first;
       pii q = g.edge_id[j].second;
@@ -97,6 +97,3 @@ std::vector<bigint> ELGMain(Graph &g){
   }
   return ans;
 }
-
-
-
